+++ spec/rubocop/cop/style/numeric_predicate_spec.rb
@@ -7,11 +7,13 @@ RSpec.describe RuboCop::Cop::Style::NumericPredicate, :config do
     inspect_source(source)
   end
 
-  shared_examples 'code with offense' do |code, expected|
+  shared_examples(
+    'code with offense'
+  ) do |code, expected:, use: expected, instead_of: code|
     context "when checking #{code}" do
       let(:source) { code }
 
-      let(:message) { "Use `#{expected}` instead of `#{code}`." }
+      let(:message) { "Use `#{use}` instead of `#{instead_of}`." }
 
       it 'registers an offense' do
         expect(cop.offenses.size).to eq(1)
@@ -46,20 +48,20 @@ RSpec.describe RuboCop::Cop::Style::NumericPredicate, :config do
     context 'when checking if a number is zero' do
       it_behaves_like 'code with offense',
                       'number == 0',
-                      'number.zero?'
+                      expected: 'number.zero?'
 
       it_behaves_like 'code with offense',
                       '0 == number',
-                      'number.zero?'
+                      expected: 'number.zero?'
 
       context 'with a complex expression' do
         it_behaves_like 'code with offense',
                         'foo - 1 == 0',
-                        '(foo - 1).zero?'
+                        expected: '(foo - 1).zero?'
 
         it_behaves_like 'code with offense',
                         '0 == foo - 1',
-                        '(foo - 1).zero?'
+                        expected: '(foo - 1).zero?'
       end
 
       context 'when comparing against a global variable' do
@@ -99,20 +101,20 @@ RSpec.describe RuboCop::Cop::Style::NumericPredicate, :config do
       context 'when target ruby version is 2.3 or higher', :ruby23 do
         it_behaves_like 'code with offense',
                         'number > 0',
-                        'number.positive?'
+                        expected: 'number.positive?'
 
         it_behaves_like 'code with offense',
                         '0 < number',
-                        'number.positive?'
+                        expected: 'number.positive?'
 
         context 'with a complex expression' do
           it_behaves_like 'code with offense',
                           'foo - 1 > 0',
-                          '(foo - 1).positive?'
+                          expected: '(foo - 1).positive?'
 
           it_behaves_like 'code with offense',
                           '0 < foo - 1',
-                          '(foo - 1).positive?'
+                          expected: '(foo - 1).positive?'
         end
       end
 
@@ -129,20 +131,20 @@ RSpec.describe RuboCop::Cop::Style::NumericPredicate, :config do
       context 'when target ruby version is 2.3 or higher', :ruby23 do
         it_behaves_like 'code with offense',
                         'number < 0',
-                        'number.negative?'
+                        expected: 'number.negative?'
 
         it_behaves_like 'code with offense',
                         '0 > number',
-                        'number.negative?'
+                        expected: 'number.negative?'
 
         context 'with a complex expression' do
           it_behaves_like 'code with offense',
                           'foo - 1 < 0',
-                          '(foo - 1).negative?'
+                          expected: '(foo - 1).negative?'
 
           it_behaves_like 'code with offense',
                           '0 > foo - 1',
-                          '(foo - 1).negative?'
+                          expected: '(foo - 1).negative?'
         end
       end
 
@@ -164,7 +166,7 @@ RSpec.describe RuboCop::Cop::Style::NumericPredicate, :config do
     context 'when checking if a number is zero' do
       it_behaves_like 'code with offense',
                       'number.zero?',
-                      'number == 0'
+                      expected: 'number == 0'
     end
 
     context 'with checking if a number is not zero' do
@@ -175,13 +177,123 @@ RSpec.describe RuboCop::Cop::Style::NumericPredicate, :config do
     context 'when checking if a number is positive' do
       it_behaves_like 'code with offense',
                       'number.positive?',
-                      'number > 0'
+                      expected: 'number > 0'
     end
 
     context 'when checking if a number is negative' do
       it_behaves_like 'code with offense',
                       'number.negative?',
-                      'number < 0'
+                      expected: 'number < 0'
+    end
+  end
+
+  context 'when there are ignored methods' do
+    let(:cop_config) do
+      {
+        'EnforcedStyle' => 'predicate',
+        'AutoCorrect' => true,
+        'IgnoredMethods' => %w[where]
+      }
+    end
+
+    context 'in argument' do
+      context 'ignored method' do
+        context 'when checking if a number is positive' do
+          context 'when target ruby version is 2.3 or higher', :ruby23 do
+            it_behaves_like 'code without offense',
+                            'where(Sequel[:number] > 0)'
+          end
+        end
+
+        context 'when checking if a number is negative' do
+          context 'when target ruby version is 2.3 or higher', :ruby23 do
+            it_behaves_like 'code without offense',
+                            'where(Sequel[:number] < 0)'
+          end
+        end
+      end
+
+      context 'not ignored method' do
+        context 'when checking if a number is positive' do
+          context 'when target ruby version is 2.3 or higher', :ruby23 do
+            it_behaves_like 'code with offense',
+                            'exclude(number > 0)',
+                            expected: 'exclude(number.positive?)',
+                            use: 'number.positive?',
+                            instead_of: 'number > 0'
+          end
+
+          context 'when target ruby version is 2.2 or lower', :ruby22 do
+            it_behaves_like 'code without offense',
+                            'exclude(number > 0)'
+          end
+        end
+
+        context 'when checking if a number is negative' do
+          context 'when target ruby version is 2.3 or higher', :ruby23 do
+            it_behaves_like 'code with offense',
+                            'exclude(number < 0)',
+                            expected: 'exclude(number.negative?)',
+                            use: 'number.negative?',
+                            instead_of: 'number < 0'
+          end
+
+          context 'when target ruby version is 2.2 or lower', :ruby22 do
+            it_behaves_like 'code without offense',
+                            'exclude(number < 0)'
+          end
+        end
+      end
+    end
+
+    context 'in block' do
+      context 'ignored method' do
+        context 'when checking if a number is positive' do
+          context 'when target ruby version is 2.3 or higher', :ruby23 do
+            it_behaves_like 'code without offense',
+                            'where { table[number] > 0 }'
+          end
+        end
+
+        context 'when checking if a number is negative' do
+          context 'when target ruby version is 2.3 or higher', :ruby23 do
+            it_behaves_like 'code without offense',
+                            'where { table[number] < 0 }'
+          end
+        end
+      end
+
+      context 'not ignored method' do
+        context 'when checking if a number is positive' do
+          context 'when target ruby version is 2.3 or higher', :ruby23 do
+            it_behaves_like 'code with offense',
+                            'exclude { number > 0 }',
+                            expected: 'exclude { number.positive? }',
+                            use: 'number.positive?',
+                            instead_of: 'number > 0'
+          end
+
+          context 'when target ruby version is 2.2 or lower', :ruby22 do
+            it_behaves_like 'code without offense',
+                            'exclude { number > 0 }'
+          end
+        end
+
+        context 'when checking if a number is negative' do
+          context 'when target ruby version is 2.3 or higher', :ruby23 do
+            it_behaves_like 'code with offense',
+                            'exclude { number < 0 }',
+                            expected: 'exclude { number.negative? }',
+                            use: 'number.negative?',
+                            instead_of: 'number < 0'
+          end
+
+          context 'when target ruby version is 2.2 or lower', :ruby22 do
+            it_behaves_like 'code without offense',
+                            'exclude { number < 0 }'
+          end
+        end
+      end
     end
   end
 end